Aligned, Tiled Access

+++ viewshed_cl.h
@@ -39,7 +39,7 @@ extern void viewshed_cl(int devices,
                         const opencl_struct * info,
                         const float * src, float * dst,
                         int cols, int rows,
-                        int x, int y, int z,
+                        int x, int y, float z,
                         double xres, double yres);
 
 #endif
